new viem version, refactor imports and types, remove aleph zero evm from support list, add api endpoint for calculating balances with dummy usd values

+++ apps/extension/src/application/trading-copilot/commands/get-staked-balance.ts
@@ -36,7 +36,7 @@ export class GetStakedBalanceCommand extends Command<Payload, Response> {
         args: [this.payload.args],
       });
 
-      const balanceAsNumber = formatEther(balance as bigint);
+      const balanceAsNumber = formatEther(balance);
 
       return new OkResult(balanceAsNumber);
     } catch (error) {
